+++ CHANGELOG.md
@@ -22,8 +22,9 @@ CHANGELOG
 - [x] 协议版本兼容性单元测试
 - [x] 内存通道版本兼容性单元测试
 - [x] 字符串路径共享内存测试
-- [ ] 握手connection的backlog支持，定时器优化
+- [x] 握手connection的backlog支持，定时器优化
 - [ ] unit test -- no router : EN_ATBUS_ERR_ATNODE_TTL
+- [x] unit test -- max msg size : EN_ATBUS_ERR_INVALID_SIZE
 - [x] msg size -> configure: 256KB
 
 1.1.0